add row usage to chunk proof
